Adds the fp upcoming services to the base interactor. Adds strings to the fpmethods
