+++ opensrp-chw/build.gradle
@@ -233,7 +233,7 @@ android {
 }
 
 dependencies {
-    implementation('org.smartregister:opensrp-client-chw-core:1.1.21-SNAPSHOT@aar') {
+    implementation('org.smartregister:opensrp-client-chw-core:1.1.22-SNAPSHOT@aar') {
         transitive = true
     }
     implementation 'androidx.appcompat:appcompat:1.0.0'
