--- opensrp-chw/src/liberia/java/org/smartregister/chw/interactor/FpUpcomingServicesInteractorFlv.java
@@ -1,4 +1,0 @@
-package org.smartregister.chw.interactor;
-
-public class FpUpcomingServicesInteractorFlv extends DefaultUpcomingServicesInteractorFlv {
-}
